refactor(candles): failure classification, retry improvements, timeout tracking
- Add MarketDataFailureKind enum and classify_market_data_failure()
- CcxtOKXAdapter: configurable timeout, init retries with backoff
- Track timeout stats alongside rate_limit in sync use cases
- Abort remaining symbols on DatabaseUnavailableError
- Curated repo instruments_list.json as default symbol source
- Move stdlib imports to TYPE_CHECKING blocks

+++ src/candles/observability/metrics.py
@@ -7,7 +7,6 @@ atomic counters protected by asyncio.Lock.
 from __future__ import annotations
 
 import asyncio
-import math
 import random
 from collections import defaultdict
 from dataclasses import dataclass, field
@@ -39,7 +38,7 @@ class ReservoirSampling:
         if not self._samples:
             return 0.0
         ordered = sorted(self._samples)
-        idx = int(round((len(ordered) - 1) * pct / 100.0))
+        idx = round((len(ordered) - 1) * pct / 100.0)
         return ordered[min(idx, len(ordered) - 1)]
 
     @property
